Update functions.php
the functions.php file now has a new admin enqueue for wp-admin.css file. By default, that stylesheet is used to hide the “Threat found” notice in the admin top bar, and also hides the “Return to Standard Editor” button for Divi.
Included snippet in functions.php to hide BlogVault in backend (from sidebar menu and plugins list) for non FreshySites users.
Included snippet in functions.php to hide the Pressable widget in the backend Dashboard page.
Included snippet in functions.php that creates a shortcode to output the Site Title. This is now used in the new footer copyright Code module, to automatically pull in the Site Title for the company name in the copyright area.